+++ src/main/java/de/maxhenkel/voicechat/events/ClientWorldEvents.java
@@ -10,4 +10,10 @@ public class ClientWorldEvents {
             listener.run();
         }
     });
+
+    public static final Event<Runnable> JOIN_SERVER = EventFactory.createArrayBacked(Runnable.class, (listeners) -> () -> {
+        for (Runnable listener : listeners) {
+            listener.run();
+        }
+    });
 }