+++ changelog.md
@@ -1,1 +1,1 @@
-- Added leave group command
+- Updated to 1.17.1-rc2